applied mentor's advice - project structure
