--- controllers/contacts-controllers.js
@@ -1,67 +1,0 @@
-const { Contact } = require("../models/contact");
-const { HttpError } = require("../helpers");
-const { ctrlWrapper } = require("../decorators");
-
-const getAllContacts = async (req, res) => {
-  const result = await Contact.find();
-  res.json(result);
-};
-
-const getContactById = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findById(contactId);
-  if (!result) {
-    throw HttpError(404, `Contact with id: ${contactId} is not found`);
-  }
-  res.json(result);
-};
-
-const addContact = async (req, res) => {
-  const result = await Contact.create(req.body);
-  res.status(201).json(result);
-};
-
-const updateContactById = async (req, res, next) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
-  if (!result) {
-    throw HttpError(404, `Contact with id: ${contactId} is not found`);
-  }
-  res.json(result);
-};
-
-const updateStatusContact = async (req, res, next) => {
-  const { contactId } = req.params;
-
-  if (Object.keys(req.body).length === 0) {
-    throw HttpError(400, "missing field favorite");
-  }
-
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
-  if (!result) {
-    throw HttpError(404, `Contact with id: ${contactId} is not found`);
-  }
-  res.json(result);
-};
-
-const removeContactById = async (req, res, next) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndDelete(contactId);
-  if (!result) {
-    throw HttpError(404, `Contact with id: ${contactId} is not found`);
-  }
-  res.status(204).send();
-};
-
-module.exports = {
-  getAllContacts: ctrlWrapper(getAllContacts),
-  getContactById: ctrlWrapper(getContactById),
-  addContact: ctrlWrapper(addContact),
-  updateContactById: ctrlWrapper(updateContactById),
-  updateStatusContact: ctrlWrapper(updateStatusContact),
-  removeContactById: ctrlWrapper(removeContactById),
-};
