+++ controllers/contacts/index.js
@@ -1,0 +1,16 @@
+const ctrlWrapper = require("../../decorators/ctrlWrapper");
+const getAllContacts = require("./getAllContacts");
+const getContactById = require("./getContactById");
+const addContact = require("./addContact");
+const updateContactById = require("./updateContactById");
+const updateStatusContact = require("./updateStatusContact");
+const removeContactById = require("./removeContactById");
+
+module.exports = {
+  getAllContacts: ctrlWrapper(getAllContacts),
+  getContactById: ctrlWrapper(getContactById),
+  addContact: ctrlWrapper(addContact),
+  updateContactById: ctrlWrapper(updateContactById),
+  updateStatusContact: ctrlWrapper(updateStatusContact),
+  removeContactById: ctrlWrapper(removeContactById),
+};
