+++ controllers/contacts/updateStatusContact.js
@@ -1,0 +1,20 @@
+const { Contact } = require("../../models");
+const { HttpError } = require("../../helpers");
+
+const updateStatusContact = async (req, res, next) => {
+  const { contactId } = req.params;
+
+  if (Object.keys(req.body).length === 0) {
+    throw HttpError(400, "missing field favorite");
+  }
+
+  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
+    new: true,
+  });
+  if (!result) {
+    throw HttpError(404, `Contact with id: ${contactId} is not found`);
+  }
+  res.json(result);
+};
+
+module.exports = updateStatusContact;
