+++ controllers/index.js
@@ -1,0 +1,5 @@
+const contactsControllers = require("./contacts");
+
+module.exports = {
+  contactsControllers,
+};
