+++ models/index.js
@@ -1,0 +1,11 @@
+const {
+  Contact,
+  addContactSchemaJoi,
+  updateFavoriteSchemaJoi,
+} = require("./contact");
+
+module.exports = {
+  Contact,
+  addContactSchemaJoi,
+  updateFavoriteSchemaJoi,
+};
