+++ routes/api/contacts.js
@@ -1,7 +1,10 @@
 const express = require("express");
 const router = express.Router();
-const contactsControllers = require("../../controllers/contacts-controllers");
-const schemas = require("../../models/contact");
+const { contactsControllers } = require("../../controllers");
+const {
+  addContactSchemaJoi,
+  updateFavoriteSchemaJoi,
+} = require("../../models");
 const { validateBody } = require("../../decorators");
 const { isValidId } = require("../../middlewares");
 
@@ -11,21 +14,21 @@ router.get("/:contactId", isValidId, contactsControllers.getContactById);
 
 router.post(
   "/",
-  validateBody(schemas.schemasJoi.addContactSchemaJoi),
+  validateBody(addContactSchemaJoi),
   contactsControllers.addContact
 );
 
 router.put(
   "/:contactId",
   isValidId,
-  validateBody(schemas.schemasJoi.addContactSchemaJoi),
+  validateBody(addContactSchemaJoi),
   contactsControllers.updateContactById
 );
 
 router.patch(
   "/:contactId/favorite",
   isValidId,
-  validateBody(schemas.schemasJoi.updateFavoriteSchemaJoi),
+  validateBody(updateFavoriteSchemaJoi),
   contactsControllers.updateStatusContact
 );
